Fix du-sound size when importing from header only
s_instr_header.full_size is 2 bytes greater for Dream 5000 than for Dream 3000, because of some additional data in IPs.
From the du-object point of view, we don't know and don't care if a du-sound binary was in format 3000 or 5000, so we have to return a unique size for both formats.
The size is usually calculated from the samples' sizes, but when we import a du-sound from header only, we only have s_instr_header.full_size to determine the size.